JDK-8304163: Move jdk.internal.module.ModuleInfoWriter to the test library
