+++ test/lib/jdk/test/lib/util/ModuleInfoWriter.java
@@ -1,12 +1,10 @@
 /*
- * Copyright (c) 2015, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2015, 2023, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
  * under the terms of the GNU General Public License version 2 only, as
- * published by the Free Software Foundation.  Oracle designates this
- * particular file as subject to the "Classpath" exception as provided
- * by Oracle in the LICENSE file that accompanied this code.
+ * published by the Free Software Foundation.
  *
  * This code is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
@@ -22,7 +20,8 @@
  * or visit www.oracle.com if you need additional information or have any
  * questions.
  */
-package jdk.internal.module;
+
+package jdk.test.lib.util;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -42,6 +41,8 @@ import jdk.internal.classfile.attribute.ModuleResolutionAttribute;
 import jdk.internal.classfile.attribute.ModuleRequireInfo;
 import jdk.internal.classfile.attribute.ModuleTargetAttribute;
 import jdk.internal.classfile.constantpool.ModuleEntry;
+import jdk.internal.module.ModuleResolution;
+import jdk.internal.module.ModuleTarget;
 
 /**
  * Utility class to write a ModuleDescriptor as a module-info.class.
